Refactored User.rb to include many modules
