+++ app/models/name.rb
@@ -1,0 +1,7 @@
+module Name
+
+  def name
+    @name = [first_name, last_name].join(" ")
+  end
+
+end
