+++ app/models/puttingaverage.rb
@@ -1,0 +1,11 @@
+module Puttingaverage
+
+  def putting_average
+    rounds.average(:putts)
+  end
+
+  def putting_average_rounded
+    putting_average.round(2)
+  end
+
+end
