+++ app/models/scoringaverage.rb
@@ -1,0 +1,11 @@
+module Scoringaverage
+
+  def scoring_average
+    rounds.average(:score)
+  end
+
+  def scoring_average_rounded
+    scoring_average.round(2)
+  end
+
+end
