--- app/models/stats.rb
@@ -1,2 +1,0 @@
-class Stats
-end
