+++ app/views/stats/index.html.erb
@@ -1,38 +1,68 @@
+<!-- <div id="myCarousel" class="carousel slide" data-ride="carousel">
+  <!-- Indicators -->
+  <!-- <ol class="carousel-indicators">
+    <li data-target="#myCarousel" data-slide-to="0" class="active"></li>
+    <li data-target="#myCarousel" data-slide-to="1"></li>
+    <li data-target="#myCarousel" data-slide-to="2"></li>
+    <li data-target="#myCarousel" data-slide-to="3"></li>
+  </ol> -->
+
+  <!-- Wrapper for slides -->
+  <!-- <div class="carousel-inner">
+    <div class="item active">
+      <%#= image_tag 'whistling.jpg', class: 'img-responsive' %>
+      <div class="carousel-caption">
+        Whistling Straits
+      </div>
+    </div> -->
+
+    <!-- <div class="item">
+      <%#= image_tag 'wekopa.jpg', class: 'img-responsive' %>
+      <div class="carousel-caption">
+        Wekopa GC
+      </div>
+    </div>
+  </div> -->
+
+  <!-- Left and right controls -->
+  <!-- <a class="left carousel-control" href="#myCarousel" role="button" data-slide="prev">
+    <span class="glyphicon glyphicon-chevron-left" aria-hidden="true"></span>
+    <span class="sr-only">Previous</span>
+  </a>
+  <a class="right carousel-control" href="#myCarousel" role="button" data-slide="next">
+    <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
+    <span class="sr-only">Next</span>
+  </a>
+</div> -->
 
 <div class="responsive-scale"></div>
-<h3 class="text-center">Saguaro  |  Wekopa - We-Ko-Pa Golf Club</h3>
+<!-- <h3 class="text-center">Saguaro  |  Wekopa - We-Ko-Pa Golf Club</h3> -->
 
-<div class="leaderboard col-xs-10 col-xs-offset-1">
-  <h2 class="text-center">Recent Scores</h2>
-    <hr />
-    <div class="col-xs-4 text-center">
-      <h3>Name</h3>
-      --------------
-      <br />
+<% if user_signed_in? %>
+  <h4 class="text-center">Welcome, <%= current_user.name %>!</h4>
+  <% end %>
+
+
+    <div class="first-chart col-lg-8 col-lg-offset-2">
+        <h3> - Latest Scores - </h3>
+      <table>
+        <tr class="table-title">
+          <th>Name</th>
+          <th>Course</th>
+          <th class="scores">Score</th>
+        </tr>
       <% @rounds.each do |round| %>
-      <h4><%= round.user.name %></h4>
+        <tr class="text-center">
+          <td><%= round.user.name %></td>
+          <td><%= round.course %></td>
+          <% if user_signed_in? %>
+          <td class="scores round-link"><%= link_to round.score, round_path(round) %></td>
+          <% else %>
+          <td class="scores"><%= round.score %></td>
+          <% end %>
+        </tr>
       <% end %>
-    </div>
-    <div class="col-xs-4 text-center">
-      <h3>Course</h3>
-      --------------
-      <br />
-      <% @rounds.each do |round| %>
-      <h4><%= round.course %></h4>
-      <% end %>
-    </div>
-    <div class="col-xs-4 text-center">
-      <h3>Score</h3>
-      --------------
-      <br />
-      <% @rounds.each do |round| %>
-      <% if user_signed_in? %>
-      <h4 class="round-link"><%= link_to round.score, round_path(round) %></h4>
-      <% else %>
-      <h4><%= round.score %></h4>
-      <% end %>
-      <% end %>
-      </div>
+      </table>
     </div>
